requirements_ml.txt
# Forex Trading System - ML Dependencies
# Install with: pip install -r requirements_ml.txt

# Core ML Frameworks
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Time-Series Forecasting
torch>=2.0.0
transformers>=4.30.0

# Microsoft Qlib (Quant Research)
pyqlib>=0.9.0

# FinRL (Deep Reinforcement Learning)
finrl>=0.3.5
stable-baselines3>=2.0.0
gym>=0.21.0

# Chinese Quant Frameworks
quantaxis>=1.10.0
# vnpy>=3.0.0  # Uncomment when ready to install

# Technical Analysis
ta-lib>=0.4.0  # May need system install first
pandas-ta>=0.3.14b0

# Backtesting
bt>=0.2.9
vectorbt>=0.25.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Performance Metrics
empyrical>=0.5.5

# API Clients
ib_insync>=0.9.86  # Interactive Brokers
v20>=3.0.25  # OANDA
requests>=2.28.0

# Utilities
python-dotenv>=1.0.0
pytz>=2023.3
tqdm>=4.65.0

# Optional GPU Acceleration
# torch-cuda>=2.0.0  # Uncomment for CUDA support
# cupy>=12.0.0  # Uncomment for GPU arrays

# Development
jupyter>=1.0.0
ipython>=8.12.0
